+++ src/main/java/br/com/fiap/challenge/dto/request/TendenciasGastosRequest.java
@@ -5,11 +5,11 @@ import jakarta.validation.constraints.NotNull;
 
 public record TendenciasGastosRequest(
         @NotNull(message = "O ano é obrigatório")
-        int ano,
+        Long ano,
         @NotNull(message = "O gasto em marketing é obrigatório")
-        double gastoMarketing,
+        Double gastoMarketing,
         @NotNull(message = "O gasto em automação é obrigatório")
-        double gastoAutomacao,
+        Double gastoAutomacao,
         @Valid
         @NotNull(message = "O id é obrigatório")
         AbstractRequest empresa
